rename chunk_strategy to abstract_chunk_strategy
