+++ src/database_builder_libs/utility/chunk/n_points_section.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from typing import Sequence
  
 from database_builder_libs.models.chunk import Chunk
-from database_builder_libs.models.chunk_strategy import AbstractChunkingStrategy, RawSection
+from database_builder_libs.models.abstract_chunk_strategy import AbstractChunkingStrategy, RawSection
 
 @dataclass(slots=True)
 class SectionChunkingStrategy(AbstractChunkingStrategy):
